backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-core
langchain-groq
